move LibraryMenuItems into its own component as its best to keep one comp per file
